refactors directory structure.
Now we have the folder "GMGridView" with the grid classes + an xcodeproj to build.
Next, there is "Example" which uses the new GMGridView.xcodeproj to link against the project.
This matches style with most other popular open source projects, and makes it easier to see what's the actual source code and what's the example. This was too mixed up previously.