allow more control over animations.
This change allows to set if an insert/update/etc should be animated or not.
It also allows adding a "scroll to action" with an extra animation flag.

Also, per default, property changes *don't* animate. We need more control here,
and setting a property never should automate *per default*. We can manually start
a animation after setting a property with layoutSubviewsWithAnimation, as seen in the updated example project.